data collection rounding numbers

+++ database/templatetags/tags.py
@@ -120,7 +120,7 @@ def get_course_stats(form, total):
     for course in courses:
         stats_list.append({
             'course': course,
-            'count': form.entries.filter(courses=course).count()/(total*1.00)*100
+            'count':int( form.entries.filter(courses=course).count()/(total*1.00)*100)
         })
 
     return stats_list
